refactor codebase and add convert dialog
