+++ __main__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+import nepcalapplet
+
+nepcalapplet.main()
